pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "vendetta"
description = "Anonymize CSV file(s)"
version = "0.0.1"
license = "MIT"

authors = []

readme = "README.md"

repository = "https://github.com/anatoly-scherbakov/vendetta"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
vendetta = "vendetta.cli:main"

[tool.poetry.dependencies]
python = ">=3.6,<3.9"
faker = "^5.5.1"
typer = "^0.3.2"
pydantic = "^1.7.3"
pyyaml = "^5.3.1"
strictyaml = "^1.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.790"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.3"
nitpick = "^0.23"

safety = "^1.10"

pytest = "^6.2"
pytest-cov = "^2.10"
pytest-randomly = "^3.5"

sphinx = "^3.4"
sphinx-autodoc-typehints = "^1.11"
doc8 = "^0.8"
m2r2 = "^0.2"
tomlkit = "^0.7"
